BOOK-005: Authentication and authorisation (#9)
## 🌟 Description
These changes are aiming to add two new end-points to the API in order
to allow users sign up and log in. If the user was successfully logged
in with valid credentials, we generate a new JWT Token and send it to
the client as cookies with an expiration of `7` days.

They also contain the authorisation middleware to validate a JWT token
received in the cookies.

## ✅ Testing
Adding unit testing for both end-points and the middleware. Also,
updated the router testing to confirm the new end-points were added to
the API. I generated the mock interface for the database, so we can mock
queries and simulate the results.

Also, I testo locally using Post-Man to sign up new users and then get
logged in with their credentials.

## 🖼️  Evidence
I tested the end-point for signup and login in Postman:

![image](https://github.com/zatarain/bookshop/assets/539783/1dd10fe0-5df9-4860-8c42-c1b0a54df146)

Confirmed that the `Authorisation` cookie is generated after login:

![image](https://github.com/zatarain/bookshop/assets/539783/49743b47-44a8-44f9-8ce0-b042b5ca2d3f)

Also, put the initial dummy end-point `/books` behind the authorisation
middleware to test:

![image](https://github.com/zatarain/bookshop/assets/539783/1b3a6f97-8996-4ca8-824d-50e690280507)

---------

Signed-off-by: Ulises Tirado Zatarain <ulises.tirado@cimat.mx>